+++ herbgrind/hg_instrument.h
@@ -20,12 +20,15 @@
 #include "runtime/hg_runtime.h"
 #include "hg_types.h"
 
+// For pointer ints
+#include <stdint.h>
+
 // Some helpful macros to have around. Some of these are from FpDebug.
 #define mkU32(_n) IRExpr_Const(IRConst_U32(_n))
 #define mkU64(_n) IRExpr_Const(IRConst_U64(_n))
 
 #define addStore(sb, src_expr, dest_addr) \
-  addStmtToIRSB(sb, IRStmt_Store(ENDIAN, mkU64((ULong)dest_addr), src_expr));
+  addStmtToIRSB(sb, IRStmt_Store(ENDIAN, mkU64((uintptr_t)dest_addr), src_expr));
 
 // Instrument a single statement, adding the instrumented statements
 // to sbOut.
